# openapi/generated.openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "p4pa-citizen",
    "description" : "Api and Models",
    "version" : "0.0.1"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
  "security" : [ {
    "BearerAuth" : [ ]
  } ],
...
  "paths" : {
    "/brokers/{brokerId}/spontaneous/debt-positions" : {
      "post" : {
        "tags" : [ "DebtPosition" ],
        "summary" : "Create spontaneous debt positions",
        "operationId" : "createSpontaneousDebtPosition",
        "parameters" : [ {
          "name" : "brokerId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/DebtPositionRequestDTO"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Created",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DebtPositionResponseDTO"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid request",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DebtPositionResponseDTO"
                }
              }
            }
          },
          "403" : {
            "description" : "Forbidden",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DebtPositionResponseDTO"
                }
              }
            }
          },
          "404" : {
            "description" : "not found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DebtPositionResponseDTO"
                }
              }
            }
          },
          "409" : {
            "description" : "Conflict",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DebtPositionResponseDTO"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal server error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DebtPositionResponseDTO"
                }
              }
            }
          }
        },
        "security" : [ {
          "BearerAuth" : [ ]
        } ]
      }
    },
...
    "/brokers/{brokerId}/organizations/{organizationId}/debt-positions/notice" : {
      "get" : {
        "tags" : [ "DebtPosition" ],
        "operationId" : "getPaymentNotice",
        "parameters" : [ {
          "name" : "X-fiscal-code",
          "in" : "header",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "brokerId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "organizationId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "installmentId",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "iuv",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "iud",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Ok",
            "content" : {
              "application/octet-stream" : {
                "schema" : {
                  "type" : "string",
                  "format" : "binary"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid request",
            "content" : {
              "application/octet-stream" : {
                "schema" : {
                  "type" : "string",
                  "format" : "binary"
                }
              }
            }
          },
          "401" : {
            "description" : "Unauthorized - Missing or invalid token",
            "content" : {
              "application/octet-stream" : {
                "schema" : {
                  "type" : "string",
                  "format" : "binary"
                }
              }
            }
          },
          "403" : {
            "description" : "Forbidden",
            "content" : {
              "application/octet-stream" : {
                "schema" : {
                  "type" : "string",
                  "format" : "binary"
                }
              }
            }
          },
          "404" : {
            "description" : "Not Found",
            "content" : {
              "application/octet-stream" : {
                "schema" : {
                  "type" : "string",
                  "format" : "binary"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal server error",
            "content" : {
              "application/octet-stream" : {
                "schema" : {
                  "type" : "string",
                  "format" : "binary"
                }
              }
            }
          }
        },
        "security" : [ {
          "BearerAuth" : [ ]
        } ]
      }
    },
...
    "/brokers/{brokerId}/debt-positions/{debtPositionId}" : {
      "get" : {
        "tags" : [ "DebtPosition" ],
        "summary" : "Retrieve a specific DebtPosition",
        "operationId" : "getDebtPositionDetail",
        "parameters" : [ {
          "name" : "brokerId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "debtPositionId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "X-fiscal-code",
          "in" : "header",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DebtPositionDTO"
                }
              }
            }
          },
          "401" : {
            "description" : "Unauthorized - Missing or invalid token",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DebtPositionDTO"
                }
              }
            }
          },
          "403" : {
            "description" : "Forbidden - Access denied",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DebtPositionDTO"
                }
              }
            }
          },
          "404" : {
            "description" : "Not Found - DebtPosition not found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DebtPositionDTO"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal Server Error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DebtPositionDTO"
                }
              }
            }
          }
        },
        "security" : [ {
          "BearerAuth" : [ ]
        } ]
      }
    },
    "/brokers/{brokerId}/debt-positions/{debtPositionId}/notice/zip" : {
      "get" : {
        "tags" : [ "DebtPosition" ],
        "summary" : "Retrieve and generate a ZIP file containing the payment notices of a specific DebtPosition",
        "operationId" : "getUnpaidPaymentNoticeZip",
        "parameters" : [ {
          "name" : "brokerId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "X-fiscal-code",
          "in" : "header",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "debtPositionId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "ZIP file containing the payment notices",
            "content" : {
              "application/octet-stream" : {
                "schema" : {
                  "type" : "string",
                  "format" : "binary"
                }
              }
            }
          },
          "204" : {
            "description" : "No Content - No installments found",
            "content" : {
              "application/octet-stream" : {
                "schema" : {
                  "type" : "string",
                  "format" : "binary"
                }
              }
            }
          },
          "401" : {
            "description" : "Unauthorized - Missing or invalid token",
            "content" : {
              "application/octet-stream" : {
                "schema" : {
                  "type" : "string",
                  "format" : "binary"
                }
              }
            }
          },
          "403" : {
            "description" : "Forbidden",
            "content" : {
              "application/octet-stream" : {
                "schema" : {
                  "type" : "string",
                  "format" : "binary"
                }
              }
            }
          },
          "404" : {
            "description" : "Not Found",
            "content" : {
              "application/octet-stream" : {
                "schema" : {
                  "type" : "string",
                  "format" : "binary"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal Server Error",
            "content" : {
              "application/octet-stream" : {
                "schema" : {
                  "type" : "string",
                  "format" : "binary"
                }
              }
            }
          }
        },
        "security" : [ {
          "BearerAuth" : [ ]
        } ]
      }
    }
  },
  "components" : {
    "schemas" : {
      "DebtPositionResponseDTO" : {
        "type" : "object",
        "properties" : {
          "debtPositionId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "organizationId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "orgFiscalCode" : {
            "type" : "string"
          },
          "orgName" : {
            "type" : "string"
          },
          "paymentDetails" : {
            "$ref" : "#/components/schemas/PaymentDetailsDTO"
          }
        },
        "required" : [ "debtPositionId", "orgFiscalCode", "orgName", "organizationId", "paymentDetails" ]
      },
      "PaymentDetailsDTO" : {
        "type" : "object",
        "properties" : {
          "iupdPagopa" : {
            "type" : "string"
          },
          "iud" : {
            "type" : "string"
          },
          "iuv" : {
            "type" : "string"
          },
          "nav" : {
            "type" : "string"
          },
          "dueDate" : {
            "type" : "string",
            "format" : "date"
          },
          "notificationFeeCents" : {
            "type" : "integer",
            "format" : "int64"
          },
          "amountCents" : {
            "type" : "integer",
            "format" : "int64"
          },
          "remittanceInformation" : {
            "type" : "string"
          }
        },
        "required" : [ "amountCents", "remittanceInformation" ]
      },
      "DebtPositionRequestDTO" : {
        "type" : "object",
        "properties" : {
          "organizationId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "debtPositionTypeOrgId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "paymentOptions" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/PaymentOptionRequestDTO"
            }
          }
        },
        "required" : [ "debtPositionTypeOrgId", "organizationId", "paymentOptions" ]
      },
      "InstallmentRequestDTO" : {
        "type" : "object",
        "properties" : {
          "notificationFeeCents" : {
            "type" : "integer",
            "format" : "int64"
          },
          "amountCents" : {
            "type" : "integer",
            "format" : "int64"
          },
          "remittanceInformation" : {
            "type" : "string"
          },
          "debtor" : {
            "$ref" : "#/components/schemas/PersonDTO"
          }
        },
        "required" : [ "amountCents", "debtor", "remittanceInformation" ]
      },
      "PaymentOptionRequestDTO" : {
        "type" : "object",
        "properties" : {
          "totalAmountCents" : {
            "type" : "integer",
            "format" : "int64"
          },
          "installments" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/InstallmentRequestDTO"
            }
          }
        },
        "required" : [ "installments", "totalAmountCents" ]
      },
      "PersonDTO" : {
        "type" : "object",
        "properties" : {
          "entityType" : {
            "$ref" : "#/components/schemas/PersonEntityType"
          },
          "fiscalCode" : {
            "type" : "string"
          },
          "fullName" : {
            "type" : "string"
          },
          "address" : {
            "type" : "string"
          },
          "civic" : {
            "type" : "string"
          },
          "postalCode" : {
            "type" : "string"
          },
          "location" : {
            "type" : "string"
          },
          "province" : {
            "type" : "string"
          },
          "nation" : {
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          }
        },
        "required" : [ "entityType", "fiscalCode", "fullName" ]
      },
      "PersonEntityType" : {
        "type" : "string",
        "enum" : [ "F", "G" ]
      },
...
      "Action" : {
        "type" : "string",
        "enum" : [ "I", "M", "A" ]
      },
      "DebtPositionDTO" : {
        "type" : "object",
        "properties" : {
          "debtPositionId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "iupdOrg" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "status" : {
            "$ref" : "#/components/schemas/DebtPositionStatus"
          },
          "debtPositionOrigin" : {
            "$ref" : "#/components/schemas/DebtPositionOrigin"
          },
          "organizationId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "debtPositionTypeOrgId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "validityDate" : {
            "type" : "string",
            "format" : "date"
          },
          "flagIuvVolatile" : {
            "type" : "boolean"
          },
          "multiDebtor" : {
            "type" : "boolean"
          },
          "flagPuPagoPaPayment" : {
            "type" : "boolean"
          },
          "creationDate" : {
            "type" : "string",
            "format" : "date-time"
          },
          "updateDate" : {
            "type" : "string",
            "format" : "date-time"
          },
          "updateOperatorExternalId" : {
            "type" : "string"
          },
          "updateTraceId" : {
            "type" : "string"
          },
          "paymentOptions" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/PaymentOptionDTO"
            }
          }
        },
        "required" : [ "debtPositionOrigin", "debtPositionTypeOrgId", "flagPuPagoPaPayment", "organizationId", "paymentOptions" ]
      },
      "DebtPositionOrigin" : {
        "type" : "string",
        "enum" : [ "ORDINARY", "ORDINARY_SIL", "SPONTANEOUS", "SPONTANEOUS_SIL", "SPONTANEOUS_MIXED", "SECONDARY_ORG", "RECEIPT_FILE", "RECEIPT_PAGOPA", "REPORTING_PAGOPA" ]
      },
      "DebtPositionStatus" : {
        "type" : "string",
        "enum" : [ "TO_SYNC", "REPORTED", "PAID", "PARTIALLY_PAID", "CANCELLED", "EXPIRED", "UNPAID", "DRAFT" ]
      },
      "InstallmentDTO" : {
        "type" : "object",
        "properties" : {
          "installmentId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "paymentOptionId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "status" : {
            "$ref" : "#/components/schemas/InstallmentStatus"
          },
          "syncStatus" : {
            "$ref" : "#/components/schemas/InstallmentSyncStatus"
          },
          "iupdPagopa" : {
            "type" : "string"
          },
          "generateNotice" : {
            "type" : "boolean"
          },
          "iud" : {
            "type" : "string"
          },
          "iuv" : {
            "type" : "string"
          },
          "iur" : {
            "type" : "string"
          },
          "iuf" : {
            "type" : "string"
          },
          "nav" : {
            "type" : "string"
          },
          "iun" : {
            "type" : "string"
          },
          "dueDate" : {
            "type" : "string",
            "format" : "date"
          },
          "switchToExpired" : {
            "type" : "boolean"
          },
          "notificationFeeCents" : {
            "type" : "integer",
            "format" : "int64"
          },
          "amountCents" : {
            "type" : "integer",
            "format" : "int64"
          },
          "remittanceInformation" : {
            "type" : "string"
          },
          "balance" : {
            "type" : "string"
          },
          "legacyPaymentMetadata" : {
            "type" : "string"
          },
          "debtor" : {
            "$ref" : "#/components/schemas/PersonDTO"
          },
          "transfers" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/TransferDTO"
            }
          },
          "notificationDate" : {
            "type" : "string",
            "format" : "date-time"
          },
          "ingestionFlowFileId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "ingestionFlowFileLineNumber" : {
            "type" : "integer",
            "format" : "int64"
          },
          "ingestionFlowFileAction" : {
            "$ref" : "#/components/schemas/Action"
          },
          "sourceFlowName" : {
            "type" : "string"
          },
          "receiptId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "creationDate" : {
            "type" : "string",
            "format" : "date-time"
          },
          "updateDate" : {
            "type" : "string",
            "format" : "date-time"
          },
          "updateOperatorExternalId" : {
            "type" : "string"
          },
          "updateTraceId" : {
            "type" : "string"
          }
        },
        "required" : [ "amountCents", "debtor", "remittanceInformation" ]
      },
      "InstallmentStatus" : {
        "type" : "string",
        "enum" : [ "TO_SYNC", "REPORTED", "PAID", "CANCELLED", "INVALID", "EXPIRED", "UNPAID", "UNPAYABLE", "DRAFT" ]
      },
      "InstallmentSyncStatus" : {
        "type" : "object",
        "properties" : {
          "syncStatusFrom" : {
            "$ref" : "#/components/schemas/InstallmentStatus"
          },
          "syncStatusTo" : {
            "$ref" : "#/components/schemas/InstallmentStatus"
          },
          "syncError" : {
            "type" : "string"
          }
        },
        "required" : [ "syncStatusFrom", "syncStatusTo" ]
      },
      "PaymentOptionDTO" : {
        "type" : "object",
        "properties" : {
          "paymentOptionId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "debtPositionId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "totalAmountCents" : {
            "type" : "integer",
            "format" : "int64"
          },
          "status" : {
            "$ref" : "#/components/schemas/PaymentOptionStatus"
          },
          "description" : {
            "type" : "string"
          },
          "paymentOptionType" : {
            "$ref" : "#/components/schemas/PaymentOptionTypeEnum"
          },
          "paymentOptionIndex" : {
            "type" : "integer",
            "format" : "int32"
          },
          "creationDate" : {
            "type" : "string",
            "format" : "date-time"
          },
          "updateDate" : {
            "type" : "string",
            "format" : "date-time"
          },
          "updateOperatorExternalId" : {
            "type" : "string"
          },
          "updateTraceId" : {
            "type" : "string"
          },
          "installments" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/InstallmentDTO"
            }
          }
        },
        "required" : [ "installments", "paymentOptionType", "totalAmountCents" ]
      },
      "PaymentOptionStatus" : {
        "type" : "string",
        "enum" : [ "TO_SYNC", "REPORTED", "PAID", "PARTIALLY_PAID", "CANCELLED", "INVALID", "EXPIRED", "UNPAID", "UNPAYABLE", "DRAFT" ]
      },
      "PaymentOptionTypeEnum" : {
        "type" : "string",
        "enum" : [ "SINGLE_INSTALLMENT", "REDUCED_SINGLE_INSTALLMENT", "INSTALLMENTS", "DOWN_PAYMENT" ]
      },
      "TransferDTO" : {
        "type" : "object",
        "properties" : {
          "transferId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "installmentId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "orgFiscalCode" : {
            "type" : "string"
          },
          "orgName" : {
            "type" : "string"
          },
          "amountCents" : {
            "type" : "integer",
            "format" : "int64"
          },
          "remittanceInformation" : {
            "type" : "string"
          },
          "stampType" : {
            "type" : "string"
          },
          "stampHashDocument" : {
            "type" : "string"
          },
          "stampProvincialResidence" : {
            "type" : "string"
          },
          "iban" : {
            "type" : "string"
          },
          "postalIban" : {
            "type" : "string"
          },
          "category" : {
            "type" : "string"
          },
          "transferIndex" : {
            "type" : "integer",
            "format" : "int32"
          },
          "creationDate" : {
            "type" : "string",
            "format" : "date-time"
          },
          "updateDate" : {
            "type" : "string",
            "format" : "date-time"
          },
          "updateOperatorExternalId" : {
            "type" : "string"
          },
          "updateTraceId" : {
            "type" : "string"
          },
          "mbdAttachment" : {
            "type" : "string"
          }
        },
        "required" : [ "amountCents", "category", "orgFiscalCode", "orgName", "remittanceInformation", "transferIndex" ]
      }
    },
    "securitySchemes" : {
      "BearerAuth" : {
        "type" : "http",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}